move pyspark dependency to separate file
